描画範囲を操作する部分の機能を修正

+++ Src/StageSelectFrame/StageSelectFrameDisplay/StageSelectDisplay.h
@@ -44,7 +44,7 @@ private:
 
 	SelectSpriteType selectSpriteType = SelectSpriteType::StagePanel;
 
-	int drawNowSelect = 0;
+	unsigned long drawNowSelect = 0;
 
 	ImageSprite stageSelectButton[ChStd::EnumCast(StageSelectButtonType::None)];
 	ChD3D11::Texture11 buttonSelectImage;
